All Done

+++ src/app/view/layout/product/product-manufacturing-cost-history/product-manufacturing-cost-history.component.ts
@@ -1,18 +1,16 @@
 import { HttpStatusCode } from '@angular/common/http';
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild, ViewEncapsulation } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 import { ActivatedRoute, Router } from '@angular/router';
-import { JwtHelperService } from '@auth0/angular-jwt';
+import { format } from 'date-fns';
 import { ToastrService } from 'ngx-toastr';
 import { catchError, map, startWith, switchMap } from 'rxjs';
 import { HttpResponse } from 'src/app/common/HttpResponse';
-import { UserRole } from 'src/app/common/UserRole';
 import { ManufacturingCost } from 'src/app/model/ManufacturingCost';
 import { Product } from 'src/app/model/Product';
-import { AuthenticationService } from 'src/app/service/authentication.service';
 import { ManufacturingCostService } from 'src/app/service/manufacturing-cost.service';
 import { ProductService } from 'src/app/service/product.service';
 import { ConfirmDialogComponent } from 'src/app/view/share/confirm-dialog/confirm-dialog.component';
@@ -20,47 +18,42 @@ import { ConfirmDialogComponent } from 'src/app/view/share/confirm-dialog/confir
 @Component({
   selector: 'app-product-manufacturing-cost-history',
   templateUrl: './product-manufacturing-cost-history.component.html',
-  styleUrls: ['./product-manufacturing-cost-history.component.scss']
+  styleUrls: ['./product-manufacturing-cost-history.component.scss'],
+  encapsulation: ViewEncapsulation.None
 })
 export class ProductManufacturingCostHistoryComponent implements OnInit {
 
   @ViewChild('paginator') paginator!: MatPaginator;
   @ViewChild(MatSort) sort!: MatSort;
-  jwtHelperService = new JwtHelperService();
-  allowAction: boolean = false;
   dataSource: MatTableDataSource<ManufacturingCost> = new MatTableDataSource<ManufacturingCost>();
-  displayedColumnsForAdmin: string[] = ['index', 'cost', 'creationTimestamp', 'action'];
-  displayedColumns: string[] = ['index', 'cost', 'creationTimestamp'];
+  displayedColumns: string[] = ['index', 'cost', 'creationTimestamp', 'action'];
   pageData: any[] = [];
   pageSizes = [5, 10, 15];
   totalElements: number = 0;
-  id: number = 0;
-  productName: string = "";
+  searchedStartDate: string = "";
+  searchedEndDate: string = "";
+  startDate!: string;
+  endDate!: string;
+  productId: number = 0;
+  productDto: Product = new Product();
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
     private matDialog: MatDialog,
     private toastrService: ToastrService,
-    private authService: AuthenticationService,
     private productService: ProductService,
     private manufacturingCostService: ManufacturingCostService
   ) { }
 
   ngOnInit(): void {
-    let jwtToken = this.authService.fetchJwtToken();
-    let loginRole = this.jwtHelperService.decodeToken(jwtToken).role;
-    if (loginRole == UserRole.ADMIN) {
-      this.allowAction = true;
-    }
-
     this.route.queryParams.subscribe(params => {
-      this.id = Number(params['id']);
+      this.productId = Number(params['productId']);
     });
 
-    this.productService.fetchById(this.id).subscribe({
+    this.productService.fetchById(this.productId).subscribe({
       next: (response: Product) => {
-        this.productName = response.name;
+        this.productDto = response;
       },
       error: (error) => {
         if (error.status == HttpStatusCode.NotFound) {
@@ -72,15 +65,15 @@ export class ProductManufacturingCostHistoryComponent implements OnInit {
   }
 
   ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
     this.assignPageData();
   }
 
   assignPageData() {
-    this.dataSource.paginator = this.paginator;
     this.paginator.page.pipe(
       startWith({}),
       switchMap(() => {
-        return this.manufacturingCostService.fetchPage(this.id, this.paginator.pageIndex + 1, this.paginator.pageSize)
+        return this.manufacturingCostService.fetchPage(this.productId, this.searchedStartDate, this.searchedEndDate, this.paginator.pageIndex + 1, this.paginator.pageSize)
           .pipe(catchError(() => {
             throw new Error("Error");
           }));
@@ -100,6 +93,22 @@ export class ProductManufacturingCostHistoryComponent implements OnInit {
     });
   }
 
+  search(): void {
+    if (this.searchedStartDate) {
+      this.searchedStartDate = format(this.searchedStartDate, "yyyy-MM-dd");
+    }
+    if (this.searchedEndDate) {
+      this.searchedEndDate = format(this.searchedEndDate, "yyyy-MM-dd");
+    }
+    if ((this.searchedStartDate && this.searchedEndDate) || (!this.searchedStartDate && !this.searchedEndDate)) {
+      this.totalElements = 0;
+      this.paginator.pageIndex = 0;
+      this.paginator.length = this.totalElements;
+      this.dataSource.paginator = this.paginator;
+      this.assignPageData();
+    }
+  }
+
   delete(id: number): void {
     const dialogRef = this.matDialog.open(ConfirmDialogComponent, {
       width: '300px', data: 'delete'
@@ -107,13 +116,24 @@ export class ProductManufacturingCostHistoryComponent implements OnInit {
 
     dialogRef.afterClosed().subscribe(result => {
       if (result) {
+        if (this.pageData.length == 1) {
+          this.toastrService.warning("There must be at least one manufacturing cost, so it cannot be deleted.", "Manufacturing Cost Deletion");
+          return;
+        }
+
         this.manufacturingCostService.delete(id).subscribe({
           next: (response: HttpResponse) => {
+            if (this.pageData.length == 1 && this.paginator.pageIndex != 0) {
+              this.paginator.pageIndex--;
+            }
+
             this.assignPageData();
             this.toastrService.success(response.message, response.title);
           },
           error: (error) => {
-
+            if (error.status == HttpStatusCode.NotFound) {
+              this.toastrService.error(error.error.message, error.error.title);
+            }
           }
         });
       }
